Restructure directories and components for the Admin Dashboard
