+++ app/javascript/admin/dashboard/components/MentorsSection.vue
@@ -1,0 +1,16 @@
+<template>
+  <div class="tab-content" id="mentors">
+    <h3>Mentors <span>2</span></h3>
+
+    <p>[VUE CHARTS GO HERE]</p>
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'admin-dashboard-section-mentors',
+}
+<script>
+
+<style scoped>
+</style>
