+++ app/javascript/admin/dashboard/components/StudentSection.vue
@@ -1,0 +1,16 @@
+<template>
+  <div class="tab-content" id="students">
+    <h3>Students <span>2</span></h3>
+
+    <p>[VUE CHARTS GO HERE]</p>
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'admin-dashboard-section-student',
+}
+<script>
+
+<style scoped>
+</style>
