+++ app/javascript/admin/dashboard/routes/index.js
@@ -1,0 +1,14 @@
+import VueRouter from 'vue-router'
+
+import Mentors from '../components/MentorsSection'
+import Students from '../components/StudentSection'
+
+export const routes = [
+  { path: '/', redirect: { name: 'students' }},
+  { path: '/students', name: 'students', component: Students },
+  { path: '/mentors', name: 'mentors', component: Mentors }
+]
+
+export const router = new VueRouter({
+  routes,
+})
