+++ app/javascript/packs/ra.js
@@ -12,6 +12,6 @@ Vue.use(TurbolinksAdapter);
 Vue.use(VTooltip)
 Vue.use(Vue2Filters)
 
-import '../admin/dashboard'
+import '../admin/dashboard/index'
 import '../ra/events'
 import '../datagrids/scores'
