+++ app/views/admin/dashboards/show.html.erb
@@ -64,6 +64,4 @@
   </div>
 </div>
 
-<div class="panel">
-  <div id="admin-dashboard"></div>
-</div>
+<div id="admin-dashboard"></div>
